# auth-service/docs/swagger.yaml
definitions:
  dto.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  dto.LoginRequest:
    properties:
      device_name:
        type: string
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  dto.ResendOTPRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  dto.ResetPasswordRequest:
    properties:
      new_password:
        maxLength: 72
        minLength: 8
        type: string
      reset_token:
        type: string
    required:
    - new_password
    - reset_token
    type: object
  dto.SessionResponse:
    properties:
      browser:
        type: string
      created_at:
        type: string
      device_name:
        type: string
      device_type:
        type: string
      expires_at:
        type: string
      id:
        type: string
      ip_address:
        type: string
      last_activity:
        type: string
      os:
        type: string
    type: object
  dto.SignupRequest:
    properties:
      email:
        maxLength: 255
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
      role:
        enum:
        - TENANT
        - LANDLORD
        type: string
    required:
    - email
    - password
    - role
    type: object
  dto.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
  dto.VerifyOTPRequest:
    properties:
      email:
        type: string
      otp:
        type: string
    required:
    - email
    - otp
    type: object
  dto.VerifyRecoveryOTPRequest:
    properties:
      email:
        type: string
      otp:
        type: string
    required:
    - email
    - otp
    type: object
  response.Response:
    properties:
      code:
        type: string
      data: {}
      message:
        type: string
      request_id:
        type: string
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Forgot password
      tags:
      - auth
  /auth/forgot-password/resend:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Resend forgot password OTP
      tags:
      - auth
  /auth/forgot-password/verify:
    post:
      consumes:
      - application/json
      parameters:
      - description: Recovery OTP
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyRecoveryOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
      summary: Verify recovery OTP
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login for TENANT, LANDLORD, or ADMIN
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.TokenResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: Login
      tags:
      - auth
  /auth/logout:
    post:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Logout current session
      tags:
      - sessions
  /auth/logout/all:
    post:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Logout from all devices
      tags:
      - sessions
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.TokenResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: Refresh access token
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Reset password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Reset password
      tags:
      - auth
  /auth/sessions:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.SessionResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get all active sessions
      tags:
      - sessions
  /auth/sessions/{id}/revoke:
    delete:
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Revoke specific session
      tags:
      - sessions
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Start signup process — sends OTP to email
      parameters:
      - description: Signup request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SignupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                message:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Email already exists
          schema:
            $ref: '#/definitions/response.Response'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/response.Response'
      summary: Initiate signup
      tags:
      - auth
  /auth/signup/resend:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ResendOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Resend signup OTP
      tags:
      - auth
  /auth/signup/verify:
    post:
      consumes:
      - application/json
      description: Verify OTP and create account
      parameters:
      - description: Verify OTP request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                message:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "410":
          description: OTP expired
          schema:
            $ref: '#/definitions/response.Response'
      summary: Verify signup OTP
      tags:
      - auth
swagger: "2.0"